+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- `v0.0.3` Grid layout
+  - Add support for ARIA role
+
+- `v0.0.5` ComboBox component
+  - Add support for grid display when `columns` prop is provided
+
 ## 0.22.0 - 2024-07-29
 ### Added
 - `v0.0.1` Callout component
@@ -180,14 +188,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## 0.11.0 - 2024-06-19
 ### Added
 - `v0.0.2` Loading component
-  
+
   - animation type `ProgressBar`
   - `className` & `TestId` props
 - `v0.0.1` SmallLoader component
 
 ### Fixed
 - `v0.0.4` NumberInput component
-  
+
   - prevent controls from changing value when input is `disabled`
 - `v0.0.4` StateIndicator component
   - styles: set overflow to hidden to disable scrollbar inside the component in Firefox